+++ mahjong/src/main/java/com/code/server/game/mahjong/logic/PlayerCardsInfoJC.java
@@ -1,6 +1,7 @@
 package com.code.server.game.mahjong.logic;
 
-import com.byz.mj.util.*;
+
+import com.code.server.game.mahjong.util.*;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -176,7 +177,7 @@ public class PlayerCardsInfoJC extends PlayerCardsInfo {
 
 	//杠牌分数计算
 	@Override
-	public void gangCompute(RoomInfo room, GameInfo gameInfo, boolean isMing, int diangangUser, String card) {
+	public void gangCompute(RoomInfo room, GameInfo gameInfo, boolean isMing, long diangangUser, String card) {
 		super.gangCompute(room, gameInfo, isMing, diangangUser, card);
 		if (!isMing) {//暗杠
 			for (Integer i : gameInfo.getPlayerCardsInfos().keySet()) {
@@ -214,7 +215,7 @@ public class PlayerCardsInfoJC extends PlayerCardsInfo {
 
 		//胡牌分数计算;
 		@Override
-	public void huCompute (RoomInfo room, GameInfo gameInfo,boolean isZimo, int dianpaoUser, String card){
+	public void huCompute (RoomInfo room, GameInfo gameInfo, boolean isZimo, long dianpaoUser, String card){
 			List<String> cs = getCardsNoChiPengGang(cards);
 			List<HuCardType> huList = HuUtil.isHu(cs, this, CardTypeUtil.cardType.get(card), new HuLimit(0));
 			//设置胡牌类型
