+++ mahjong/src/main/java/com/code/server/game/mahjong/logic/PlayerCardsInfoKD.java
@@ -1,9 +1,10 @@
 package com.code.server.game.mahjong.logic;
 
-import com.byz.mj.util.CardUtil;
-import com.byz.mj.util.HuCardType;
-import com.byz.mj.util.HuLimit;
-import com.byz.mj.util.HuUtil;
+
+import com.code.server.game.mahjong.util.CardUtil;
+import com.code.server.game.mahjong.util.HuCardType;
+import com.code.server.game.mahjong.util.HuLimit;
+import com.code.server.game.mahjong.util.HuUtil;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -174,7 +175,7 @@ public class PlayerCardsInfoKD extends PlayerCardsInfo {
 
 	//杠牌分数计算
 	@Override
-    public void gangCompute(RoomInfo room,GameInfo gameInfo,boolean isMing,int diangangUser,String card){
+    public void gangCompute(RoomInfo room, GameInfo gameInfo, boolean isMing, long diangangUser, String card){
 		super.gangCompute(room,gameInfo,isMing,diangangUser,card);
 		if(!isMing){//暗杠
         	for (Integer i : gameInfo.getPlayerCardsInfos().keySet()){
@@ -216,7 +217,7 @@ public class PlayerCardsInfoKD extends PlayerCardsInfo {
 
     //胡牌分数计算
 	@Override
-    public void huCompute(RoomInfo room,GameInfo gameInfo,boolean isZimo,int dianpaoUser,String card){
+    public void huCompute(RoomInfo room, GameInfo gameInfo, boolean isZimo, long dianpaoUser, String card){
 		List<String> cs = getCardsNoChiPengGang(cards);
 		List<HuCardType> huList = HuUtil.isHu(cs, this,CardTypeUtil.cardType.get(card) , new HuLimit(0));
 		//设置胡牌类型
